[nvim]: use GH Copilot as avante provider
The Scout provider is working but I want to use a default provider for a
while to see if it works a little better. Also, there was some streaming
problem with the scout provider.

Also fixed the copilot node path to use the one installed with asdf
instead of using the one provided by the project